docs(#5) :: type 모듈화
