+++ src/app/layout.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import GlobalStyle from "@/styles/GlobalStyle";
 import styled from "@emotion/styled";
 import React from "react";
