+++ src/store/user.ts
@@ -1,17 +1,5 @@
 import { create } from "zustand";
-
-export type User = {
-  img: string;
-  name: string;
-  mail: string;
-};
-
-type UserStore = {
-  user: User;
-  setUser: (u: User) => void; // 통째로 교체
-  patchUser: (u: Partial<User>) => void; // 일부만 업데이트
-  reset: () => void; // 초기값으로 리셋
-};
+import { User, UserStore } from "@/types/types";
 
 const initialUser: User = {
   img: "https://i.pinimg.com/736x/15/b4/f3/15b4f3b63c2e43017a5c2d7d3660292e.jpg",
